Move SimpleChain source files to src/chain subdir
