+++ src/chain/model/primitives/Hash.ts
