+++ src/chain/model/primitives/HexData.ts
