+++ src/chain/model/primitives/common.ts
