+++ src/chain/model/strings.ts
