+++ src/chain/vm/SimpleVM.ts
