+++ src/chain/vm/getLatestBlock.ts
