+++ src/chain/vm/getNextBlock.ts
