+++ src/chain/vm/initializeVM.ts
